requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
aiosqlite>=0.20
httpx>=0.27
aiomqtt>=2.0
pydantic-settings>=2.2
jinja2>=3.1
python-multipart>=0.0.9
bcrypt>=4.1
websockets>=12.0
python-dotenv>=1.0
